% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcS2.R
\name{CalcS2}
\alias{CalcS2}
\alias{LevelVariance}
\alias{KonVariance}
\title{Calculation of sigma squared}
\usage{
CalcS2(
  data,
  yVar,
  xVar = NULL,
  strataVar,
  sampleVar = NULL,
  residVariance = TRUE,
  baseVar = NULL,
  estimateType = list("level", "diff", "ratio"),
  meanType = c("y", "var"),
  p = 1,
  printPlot = FALSE
)
}
\arguments{
\item{data}{The dataset}

\item{yVar}{The variable name for the interest variable (y). Eg "turnover"}

\item{xVar}{The name of the activity variable which relates to the interest
variable. Eg "numberOfEmployees"}

\item{strataVar}{The variable used for stratification}

\item{sampleVar}{The variable used to identify which companies were included
in the sample. This is used in cases where the interest variable (y
variable) are not available for the population. When a variable is
specified here, y-values from the sample only are used to calculate s2.
Default is NULL}

\item{residVariance}{Whether to calculate residual or other variance.
Default is residual.}

\item{baseVar}{Variable which provides values as a fixed base for difference
and ratio statistics. If this is missing, then a moving base is used as the
previous period.}

\item{estimateType}{The type of statistic being calculated. Choose beteen
"level" for a standard level statistic, "diff" for a difference statistic,
or "ratio" = ratio statistic. If not provided, the function defaults to a
level statistic.}

\item{meanType}{In the case of several periods of data, meanType specifies
how they should be averaged. Default "var" calculated sigma sq and takes the
average. "y" takes the average of the y values first and then calculates
sigma sq.}

\item{p}{Specifies how much overlap there is between periods for a
difference or ratio statistic. Default is p = 1 which indicates full overlap
between periods.}

\item{printPlot}{Indicates whether or not to print boxplots of the residual
values (ei or ai) to identify outliers. One plot is made for each period or
each difference/ratio period. Default is FALSE.}
}
\value{
\item{s2}{sigma squred values for each strata group} \item{N}{number
of observations used in the calculations.} \item{n}{If a sample variable is
specified, n is also returned}
}
\description{
Calculation of sigma squared using either residual or simple variance for
level, difference or ratio statistics. The function includes options for
averaging over several periods.
}
\examples{
  
# Call test dataset
  data(testData)
  
# Create stratification variable
  testData$strata1 <- paste(testData$nace3, testData$storGrp, sep="")

# Resdiual variance for one period
  CalcS2(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  estimateType = "level")

# Simple variance for one period
  CalcS2(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  estimateType = "level", residVariance = FALSE)

# Residual variance for three periods
  CalcS2(data = testData, yVar = c("y1", "y2", "y3"), xVar = "antAnsatt", strataVar = "strata1", 
  estimateType = "level") 

# Include a sample variable
  CalcS2(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  sampleVar = "utv1", estimateType = "level")
  
# Difference example - fixed base
  CalcS2(data = testData, yVar = c("y1", "y2", "y3"), xVar = "storGrp", strataVar = "strata1", 
  residVariance = TRUE, baseVar = "y1", estimateType = "diff")
  
# Difference example - changing base (just leave out baseVar)
  CalcS2(data = testData, yVar = c("y1", "y2", "y3"), xVar = "storGrp", strataVar = "strata1", 
  residVariance = TRUE, estimateType = "diff")

#Ratio example showing boxplots - fixed base
  CalcS2(data = testData, yVar = c("y1", "y2","y3"),  xVar = "antAnsatt", strataVar = "strata1",
  baseVar="y1", estimateType = "ratio", printPlot = TRUE)


}
\keyword{survey}
