% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcY.R
\name{CalcY}
\alias{CalcY}
\title{Calculation of the estimate for the interest variable}
\usage{
CalcY(
  data,
  yVar,
  xVar = NULL,
  strataVar,
  sampleVar = NULL,
  estimateType = list("level", "diff", "ratio"),
  residVariance = TRUE
)
}
\arguments{
\item{data}{The dataset}

\item{yVar}{The variable name for the interest variable (y). Eg "turnover"}

\item{xVar}{The name of the activity variable which relates to the interest
variable. This is only required for estimation with a rate model. Eg
"numberOfEmployees"}

\item{strataVar}{The variable used for stratification}

\item{sampleVar}{The variable used to identify which companies were included
in the sample. This is used in cases where the interest variable (y
variable) are not avaialable for the population. When a variable is
specified here, y-values from the sample only are used to calculate s2.
Default is NULL}

\item{estimateType}{The type of estimate to do. Default is set to 'level'
but option for 'diff' for a difference statistis and 'ratio' for a ratio
statistic.}

\item{residVariance}{Whether to calculate the total based on a rate model
(residVariance = TRUE) or not (residVariance = FALSE). This option is only
available for level statistic calculation at the moment.}
}
\value{
The estimates for y are returned as a vector of length equal to the
number of strata (identified in strataVar), or a dataset if several periods
are given with each column giving the estimates in each strata.
}
\description{
Calculation of the estimates for the interest variable (Y) using either a
rate model (with an x variable) or simple expansion. Includes options for
calculation of level, ratio and difference statistics.
}
\examples{
  
# Call test dataset
  data(testData)
  
# Create stratification variable
  testData$strata1 <- paste(testData$nace3, testData$storGrp, sep="")

# Examples for level statistic
  CalcY(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  estimateType = "level")
  
  CalcY(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  estimateType = "level", residVariance = FALSE)
  
  CalcY(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  sampleVar = "utv1", estimateType = "level")
  
  CalcY(data = testData, yVar = "y1", xVar = "antAnsatt", strataVar = "strata1", 
  sampleVar = "utv1", estimateType = "level", residVariance = FALSE)

# Example for difference statistic
  CalcS2(data = testData, yVar = c("y1", "y2", "y3"), xVar = "storGrp", strataVar = "strata1", 
  baseVar = "y1", estimateType = "diff")

# Example for ratio statistic
  CalcS2(data = testData, yVar = c("y1", "y2", "y3"), xVar = "storGrp", strataVar = "strata1", 
  baseVar = "y1", estimateType = "ratio")

}
\keyword{survey}
