% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_the_largest.R
\name{displace}
\alias{displace}
\title{A function that shifts the elements in a vector one step to the right, the last element disappears 
while x_1 is inserted as the first element}
\usage{
displace(x, x_1 = 0)
}
\arguments{
\item{x}{A vector of numbers (NA's are allowed)}

\item{x_1}{A single number (NA are allowed)}
}
\value{
A vector
}
\description{
A function that shifts the elements in a vector one step to the right, the last element disappears 
while x_1 is inserted as the first element
}
\keyword{internal}
