% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillStrata.R
\name{FSmatrix}
\alias{FSmatrix}
\alias{FStotnAlloc}
\alias{FSvarAlloc}
\alias{FScvAlloc}
\alias{FScvAlloc2}
\alias{FSvarLimits}
\alias{FStotVar}
\alias{FStotCV}
\alias{FillStrata}
\title{Optimal allocation of sample sizes}
\usage{
FSmatrix(s2, N, ...)

FStotnAlloc(s2, N, totn, ...)

FSvarAlloc(s2, N, varLimits, group, ...)

FScvAlloc(s2, N, totY, cvLimits, group, ...)

FScvAlloc2(s2, N, totYgroup, cvLimits, group, ...)

FSvarLimits(totY, cvLimits, group, ...)

FStotVar(s2, N, min_n, group, ...)

FStotCV(s2, N, totY, min_n, group, ...)

FillStrata(
  s2,
  N,
  totVar = function(s2, n, N) {
     s2 * N * (N - n)/n
 },
  nullcorrection = 1e+20,
  totn = min(sum(max_n_exact), 20000),
  min_n = NULL,
  use_n = NULL,
  corrSumVar = FALSE,
  dotprint = totn > 1000,
  varLimits = NULL,
  group = rep("All", length(s2)),
  intern = FALSE,
  max_n = N,
  totY = NULL,
  cvLimits = NULL,
  totYgroup = NULL,
  groupNames = NULL,
  returnGroupN = FALSE,
  returnGroupn = FALSE,
  returnTable = FALSE,
  returnTotYgroup = FALSE,
  returnVarLimits = FALSE,
  returnMulti = FALSE,
  returnMatrix = FALSE,
  totnGroup = NULL,
  min_nGroup = NULL
)
}
\arguments{
\item{s2}{An estimate for the variance within each strata. For example s2 =
c(35, 3, 2, 30, 7)}

\item{N}{The population size within each strata. For example N = c(100, 100,
500, 50, 150)}

\item{totn}{The total desired sample size}

\item{varLimits}{The desired maximum variance for each group.  Note: When
varLimits is given, a different output value is returned - the sample size
that fulfills the required group variance for totVar.  If varLimits = Inf,
the output is the group sum of totVar with "n = min_n".}

\item{group}{A vector or dummy matrix giving the groups to be used with
varLimits}

\item{totY}{Estimate of total Y in each stratum.}

\item{cvLimits}{The desired maximum CV for each group. This can be a single
value or a vector of values.}

\item{totYgroup}{Estimate of total Y in each group. Alternative instead of
totY.}

\item{min_n}{A minimum sample size in each stratum. This can be a single
number used in all strata or a vector of numbers (one for each stratum).}

\item{totVar}{The function for calculating the total variance from s2, n
(sample size) and N as inputs.  The function should be able to handle
vectors as inputs and give a vector as an output.}

\item{nullcorrection}{Variance at n=0 is set to nullcorrection times
variance at n=1. A large value is needed by the algorithm (not Inf).}

\item{use_n}{Matrix with the same number of columns as there are strata.
Each column should contain the desired sample size choices (will accept
missing values)}

\item{corrSumVar}{Logical value giving if the corrected variance should be
used. Default is FALSE.}

\item{dotprint}{Condition to invoke printing of "---------#--------...."
during calculations. Default is set for when totn > 1000.}

\item{intern}{Logical value giving if the function is being used internally
within this or another function or not. Default is set to FALSE.}

\item{max_n}{A maximum sample size in each stratum. This can be a single
number used in all strata or a vector of numbers (one for each stratum).}

\item{groupNames}{Alternative instead of finding group names from other
input.}

\item{returnGroupN}{Logical on whether to return the population group size.}

\item{returnGroupn}{Logical on whether to return the sample group size.}

\item{returnTable}{Logical on whether to return the table}

\item{returnTotYgroup}{Logical on whether to return the total Y values}

\item{returnVarLimits}{Logical on whether to return the  Y limits}

\item{returnMulti}{Logical on whether to return the multi}

\item{returnMatrix}{Logical on whether to return the matrix}

\item{totnGroup}{Maximal total sample size within group (override cvLimits)}

\item{min_nGroup}{Minimal total sample size within group (override cvLimits)}
}
\value{
The output from \bold{FSmatrix} (FillStrata with returnMatrix=TRUE)
is a two-dimensional list.

\item{Column:Strata}{Output values according to each strata.}
\item{Column:Group}{Output values according to each group as specified by
input.} \item{Column:All}{Output by treating all strata as a single group.}
\item{Row:n}{Allocated sample size} \item{Row:N}{Population size}
\item{Row:totVar}{Total variance} \item{Row:totY}{Total Y} \item{Row:cv}{CV
(\%)} \item{Row:varLimits}{As input (only Group)} \item{Row:cvLimits}{As
input or computed from varLimits (only Group)}

The output from \bold{FillStrata} is a list with two elements when varLimits
or cvLimits is not used.  \item{filledstrata}{Strata is filled sequentially
according to this vector. } \item{sumVar}{Total variance for all values of
total n. }

When returnTable=TRUE and when varLimits or cvLimits is used (FStotnAlloc,
FSvarAlloc, FScvAlloc, FScvAlloc2) the output is the allocated sample size
in each strata.

Other elements of the FSmatrix-output can be obtained by setting other
return-parameters to TRUE or by using varLimits=Inf or cvLimits=Inf.
}
\description{
This function finds an optimal allocation by filling up strata one unit at a
time. This approach works when the variance function of n decreases smoothly
(The derivative should be strictly increasing).
}
\examples{

# Examples with 5 strata based on FSmatrix
  s2    = c(35, 3, 2, 30, 7)         # Variance estimate
  N     = c(100, 100, 500, 50, 150)  # N in each strata
  names(N) = paste("st",1:5,sep="")  # Strata-names
  totY  = c(8800,900,5500,400,300)   # Totals in each strata
  g     = c("grA","grA","grB","grB","grB") # Two groups
 
# Total sample size of 200 chosen in advance
  m = FSmatrix(s2=s2,N=N,totn=200,group=g)
  # Three ways of getting allocated n in each strata
  m[,1]$n
  m[1,]$Strata
  m[[1,1]]
  
# Alloc from varLimits
  vL = c(5000,10000)   # variance limits for groups 
  m = FSmatrix(s2=s2,N=N,varLimits=vL,group=g)
  m[[1,1]] # Allocated n in each strata
  m[3,] # Total variance in strata, group (less than limits) and all.
  
#  Calculate CV from totY in strata from known allocation (20\%)
  m = FSmatrix(s2=s2,N=N,totY=totY,max_n=round(N/5),group=g)
  m[5,] # CV in strata, group and all.

}
\keyword{survey}
