% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{get_extreme}
\alias{get_extreme}
\title{Get extreme values
Get extreme values in the sample dataset}
\usage{
get_extreme(data, id, x, y, stratum)
}
\arguments{
\item{data}{Population data frame with additional variables from rate_model output}

\item{id}{Name of identification variable as a string. Should be same in sample and data dataframes.}

\item{x}{Name of the explanatory variable}

\item{y}{Name of the statistic variable}

\item{stratum}{Name of the stratum variable}
}
\value{

}
\description{
Get extreme values
Get extreme values in the sample dataset
}
