% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{get_groups}
\alias{get_groups}
\title{Get group estimates
Get estimates for group from rate model output}
\usage{
get_groups(data, x, y, stratum, groups)
}
\arguments{
\item{data}{population data frame with additional variables from rate_model output}

\item{x}{name of the explanatory variable}

\item{y}{name of the statistic vaiable}

\item{stratum}{name of the stratum variable}

\item{groups}{name of variable(s) for using for groups}
}
\value{

}
\description{
Get group estimates
Get estimates for group from rate model output
}
