% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{get_results}
\alias{get_results}
\title{Get formatted results
Get a table for results in each stratum for a rate model}
\usage{
get_results(data, x, y, stratum)
}
\arguments{
\item{data}{Population data frame with additional variables from rate_model output}

\item{x}{Name of the explanatory variable}

\item{y}{Name of the statistic vaiable}

\item{stratum}{Name of the stratum variable}
}
\value{

}
\description{
Get formatted results
Get a table for results in each stratum for a rate model
}
