% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{get_strata_results}
\alias{get_strata_results}
\title{Get strata results
Get a table for results in each stratum for a rate model}
\usage{
get_strata_results(data, x = NULL, y = NULL, strata = ".strata")
}
\arguments{
\item{data}{Population data frame with additional variables from rate_model output}

\item{x}{Name of the explanatory variable}

\item{y}{Name of the statistic variable}

\item{strata}{Name of the stratification variable}
}
\value{
Table with strata results
}
\description{
Get strata results
Get a table for results in each stratum for a rate model
}
