% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{get_var}
\alias{get_var}
\title{Get variable names}
\usage{
get_var(data, var)
}
\arguments{
\item{data}{Data frame to find variables in}

\item{var}{Variable to find. Can be 'x', 'y', 'id' or 'strata'}
}
\value{
variable name
}
\description{
Get variable names
}
