% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lonely_strata.R
\name{merge_lonely_strata}
\alias{merge_lonely_strata}
\title{Merge lonely strata}
\usage{
merge_lonely_strata(
  data,
  sample_data = NULL,
  id,
  x,
  y,
  strata,
  exclude = NULL,
  block = NULL,
  group = NULL
)
}
\arguments{
\item{data}{Population data frame}

\item{sample_data}{Sample data frame}

\item{id}{Name of identification variable as a string. Should be same in sample_data and data dataframes.}

\item{x}{Name of the explanatory variable.}

\item{y}{Name of the statistic variable.}

\item{strata}{Name of the strata variable(s).}

\item{exclude}{Vector of id numbers to exclude from the model. This observations will be included 
in the total with their observed values.}

\item{block}{Name of variable(s) for using for blocks within which strata merge may take place.}

\item{group}{Name of variable(s) for using for groups.}
}
\value{
Data frame for whole population with estimation strata obtained from strata merge, 
and an indicator for merge (1 if merge took place, 0 otherwise).
}
\description{
Merge lonely strata that are being neither surprise nor full-count within blocks.
}
