% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{plot_extreme}
\alias{plot_extreme}
\title{Plot extreme values
Plot comparison of the different extreme values using G-values or estimate comparisons.}
\usage{
plot_extreme(data, id, y, size = 10, type = "G", ylim = NULL)
}
\arguments{
\item{data}{Data frame output from get_extreme() function}

\item{id}{Name of identification variable as a string. Should be same in sample and data dataframes.}

\item{y}{Name of the statistic variable}

\item{size}{The number of outlier units to show. Default is 10.}

\item{type}{The type of plot to show. 'G' output a plot of the G values compared to the boundary, type 'estimate' gives a comparison in strata estimates with and without the observation.}

\item{ylim}{The upper limit to show in the plot.}
}
\value{

}
\description{
Plot extreme values
Plot comparison of the different extreme values using G-values or estimate comparisons.
}
