% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{Population test data}
\format{
A data frame with 10000 rows and 7 variables:
\describe{
  \item{id}{identification variable}
  \item{employees}{number of employees}
  \item{employees_f}{number of female employees}
  \item{employees_m}{number of male employees}
  \item{turnover}{amount of turnover of the company in NOK}
  \item{size}{category size of the company in terms of turnover (small, mid, large)}
  \item{industry_group}{the activity group for which the business works in}
}
}
\usage{
pop_data
}
\description{
Test population data for using in struktuR package (synthetic data).
}
\keyword{datasets}
