% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pop_data2}
\alias{pop_data2}
\title{Alternative population test data}
\format{
A data frame with 10000 rows and 5 variables:
\describe{
  \item{id}{Identification variable}
  \item{employees}{Number of employees}
  \item{turnover}{Amount of turnover of the company (in 000 NOK)}
  \item{size}{Category size of the company in terms of turnover (small, mid, large)}
  \item{industry_group}{The activity group for which the business works in.}
  \item{job_vacancies}{The number of job vacancies in the company for those that responded to the survey.}
  \item{sick_days}{The number of sick leave days recorded for the company for those that responded to the survey.}
}
}
\usage{
pop_data2
}
\description{
Test population data for using in struktuR package (synthetic data). This alternative version
includes two statistic variables for 1000 companies in the survey (although 5 non-response).
}
\keyword{datasets}
