% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{pop_data_1obs}
\alias{pop_data_1obs}
\title{Population test data with a 1 observation stratum}
\format{
A data frame with 10005 rows and 7 variables:
\describe{
  \item{id}{Identification variable}
  \item{employees}{Number of employees}
  \item{employees_f}{number of female employees}
  \item{employees_m}{number of male employees}
  \item{turnover}{Amount of turnover of the company (in 000 NOK)}
  \item{size}{Category size of the company in terms of turnover (small, mid, large)}
  \item{industry}{The activity group for which the business works in.}
}
}
\usage{
pop_data_1obs
}
\description{
Test population data for using in struktuR package (synthetic data).
}
\keyword{datasets}
