% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{rate_model}
\alias{rate_model}
\title{Rate model
Estimates and diagnostics for estimating a total from a rate model}
\usage{
rate_model(data, sample, id, x, y, stratum)
}
\arguments{
\item{data}{Population data frame}

\item{sample}{Sample data frame}

\item{id}{Name of identification variable as a string. Should be same in sample and data dataframes.}

\item{x}{Name of the explanatory variable}

\item{y}{Name of the statistic vaiable}

\item{stratum}{Name of the stratum variable}
}
\value{

}
\description{
Rate model
Estimates and diagnostics for estimating a total from a rate model
}
