% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{robust_var}
\alias{robust_var}
\title{Robust variance estimation
Internal function for robust estimation of variance}
\usage{
robust_var(X_pop, X_utv, ei, hi)
}
\arguments{
\item{X_utv}{Total sum of x variable in the sample}

\item{ei}{Residuals}

\item{hi}{Hat values}

\item{Xpop}{Total sum of x variable in the population}
}
\value{

}
\description{
Robust variance estimation
Internal function for robust estimation of variance
}
