% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_alloc.R
\name{round2}
\alias{round2}
\title{Rounding}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{A vector of numbers to be rounded}

\item{digits}{An integer indicating the number of decimal places (default 0). Negative values are allowed (see 'Description')}
}
\value{
The rounded numbers
}
\description{
Rounds according to the "round up" rule: When a number is halfway between two others, it is rounded toward the nearest number that is away from zero.
Rounding to a negative number of digits means rounding to a power of ten, so for example round(x, digits = -2) rounds to the nearest hundred.
}
\keyword{internal}
