% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample_data_fulltelling}
\alias{sample_data_fulltelling}
\title{Sample test data with full count stratum}
\format{
A data frame with 870 rows and variables:
\describe{
  \item{id}{Identification variable}
  \item{employees}{Number of employees}
  \item{employees_f}{Number of female employees}
  \item{employees_m}{Number of male employees}
  \item{turnover}{Amount of turnover of the company (in 000 NOK)}
  \item{size}{Category size of the company in terms of turnover (small, mid, big)}
  \item{industry}{The activity group for which the business works in}
  \item{job_vacancies}{The number of job vacancies in the company.}
  \item{sick_days}{The number of sick leave days recorded for the company.}
  \item{sick_days_f}{The number of sick leave days recorded for female employees for the company.}
  \item{sick_days_m}{The number of sick leave days recorded for male employees for the company.}
}
}
\usage{
sample_data_fulltelling
}
\description{
Test sample data for using in struktuR package. It includes synthetic
data with 870 rows, each row being a company that was selected to the survey.
}
\keyword{datasets}
