% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{struktur_model}
\alias{struktur_model}
\title{Run a Struktur model
Estimates total and uncertainty for a rate, homogeneous or regression model within strata.}
\usage{
struktur_model(data, sample_data = NULL, id, x, y, strata, method = "rate")
}
\arguments{
\item{data}{Population data frame}

\item{sample_data}{Sample data frame or variable with sample indicator (1 for in sample, 0 otherwise)}

\item{id}{Name of identification variable as a string. Should be same in sample_data and data dataframes.}

\item{x}{Name of the explanatory variable.}

\item{y}{Name of the statistic variable.}

\item{strata}{Name of the strata variable.}
}
\value{

}
\description{
Run a Struktur model
Estimates total and uncertainty for a rate, homogeneous or regression model within strata.
}
