% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{get_extremes}
\alias{get_extremes}
\title{Get extreme values
Get extreme values in the sample dataset}
\usage{
get_extremes(data, id, x, y, strata, na_rm = TRUE)
}
\arguments{
\item{data}{Population data frame with additional variables from rate_model output}

\item{id}{Name of identification variable as a string. Should be same in sample and data dataframes.}

\item{x}{Name of the explanatory variable}

\item{y}{Name of the statistic variable}

\item{strata}{Name of the stratification variable}
}
\value{

}
\description{
Get extreme values
Get extreme values in the sample dataset
}
