% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{get_groups}
\alias{get_groups}
\title{Get group estimates
Get estimates for group from rate model output}
\usage{
get_groups(data, x, y, strata, group)
}
\arguments{
\item{data}{Population data frame with additional variables from rate_model output}

\item{x}{Name of the explanatory variable.}

\item{y}{Name of the statistic variable to estimate.}

\item{strata}{Name of the stratification variable for running models within.}

\item{group}{Name of variable(s) for using for groups.}
}
\value{

}
\description{
Get group estimates
Get estimates for group from rate model output
}
