% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot CVs
Plot comparison of the different robust estimations of the CVs}
\usage{
plot_cv(data, y, strata)
}
\arguments{
\item{data}{Data which is the output from get_results() function}

\item{y}{Name of the statistic variable}

\item{strata}{Name of stratification variable.}
}
\value{

}
\description{
Plot CVs
Plot comparison of the different robust estimations of the CVs
}
