% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_model.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot CVs
Plot comparison of the different robust estimations of the CVs}
\usage{
plot_cv(results, y)
}
\arguments{
\item{results}{output from get_results() function}

\item{y}{name of the statistic vaiable}
}
\value{

}
\description{
Plot CVs
Plot comparison of the different robust estimations of the CVs
}
