% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{struktur_model}
\alias{struktur_model}
\title{Run a struktur model}
\usage{
struktur_model(
  data,
  sample_data = NULL,
  id,
  x,
  y,
  strata,
  exclude = NULL,
  method = "rate"
)
}
\arguments{
\item{data}{Population data frame}

\item{sample_data}{Sample data frame or variable with sample indicator (1 for in sample, 0 otherwise)}

\item{id}{Name of identification variable as a string. Should be same in sample_data and data dataframes.}

\item{x}{Name of the explanatory variable.}

\item{y}{Name of the statistic variable.}

\item{strata}{Name of the strata variable.}

\item{exclude}{A list of observations in the sample that should be excluded from the model. 
The observed values are included in the total estimate.}

\item{method}{The type of the model to use.Default is "rate"}
}
\value{
Data frame for whole population with mass-imputation values.
Other variables, beta estimates are also included
}
\description{
Estimates total and uncertainty for a rate, homogeneous or regression model within strata.
}
