% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struktur_model.R
\name{struktur_model}
\alias{struktur_model}
\title{Run a struktur model}
\usage{
struktur_model(
  data,
  sample_data = NULL,
  id,
  x,
  y,
  strata,
  exclude = NULL,
  method = "rate"
)
}
\arguments{
\item{data}{Population data frame}

\item{sample_data}{Sample data frame}

\item{id}{Name of identification variable as a string. Should be same in sample_data and data dataframes.}

\item{x}{Name of the explanatory variable.}

\item{y}{Name of the statistic variable.}

\item{strata}{Name of the strata variable.}

\item{exclude}{Vector of id numbers to exclude from the model. This observations will be included 
in the total with their observed values.}

\item{method}{Model methods to use. Default is 'rate' for a rate model. Also 'homogen' and 'reg'
for regression model will be available soon.}
}
\value{
Data frame for whole population with mass-imputation values.
Other variables, beta estimates are also included
}
\description{
Estimates total and uncertainty for a rate, homogeneous or regression model within strata.
}
